move added UT to COW merge unless MOR fix worked out
